requirements.txt
black==24.4.2
numpy==1.26.4
pygame==2.5.2
PyOpenGL==3.1.5
